Fix different png images getting same hash.
If png images with the same size but different
content using only black on a transparent
background the first included image replaces all
subsequent images even if different since they get
the same hash because idata string is the same.

Creating the hash from the actual content of the
file instead of idata fixes this issue.